PDF_Summarizer_Openai(nao_funcional)/requirements.txt
gradio
PyPDF2
langchain
langchain-community
langchain-openai
openai
python-dotenv
